requirements.txt
langchain-core
python-dotenv
streamlit
langchain_community
langserve
uvicorn
sse_starlette
bs4
chromadb
groq
langchain-groq
wikipedia
arxiv
langchainhub
scholarly
pubmed-parser
google-search-results>=2.4.2
fastapi
uvicorn